// src/game/types.ts
export type Suit = "oros" | "copes" | "espases" | "bastos";
export type Rank = 1 | 3 | 4 | 5 | 6 | 7;

export interface Card {
  suit: Suit;
  rank: Rank;
  /** ID únic per renderitzar (ex: "1-espases") */
  id: string;
}

export type PlayerId = 0 | 1 | 2 | 3;
// ...
export interface TrickCard {
  player: PlayerId;
  card: Card;
  /** Si la carta s'ha jugat tapada (boca avall). En aquest cas no compta
   *  per a la força de la baza ni per al càlcul de l'envit del jugador. */
  covered?: boolean;
}